requirements.txt
Django==4.2.20
django-ckeditor==6.7.2
django-js-asset==2.2.0
pillow==10.4.0
requests==2.32.3
dj-database-url==2.2.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-decouple==3.8
whitenoise==6.7.0
whitenoise==6.7.0